refactoring to avoid have v-on:click if not necessary (screen reader always tell element clickable if v-on:click is present)
